Simplified PrivateMembers
If the confidential flag is set, the Register container will use the PrivateMembers component.
However, the API does not return members if the group is confidential meaning nothing is
ever rendered.

One exception to "nothing is ever rendered" is when the the temporary/default user is
added to the list during a registration attempt.

*This* is the reason the component was kept; could probably be simplied even more if desired.